server/requirements.txt
aiohttp==3.10.11
annotated-types==0.7.0
anyio==4.11.0
click==8.3.0
fastapi==0.118.0
greenlet==3.2.4
h11==0.16.0
httptools==0.6.4
idna==3.10
psycopg2-binary==2.9.10
pydantic==2.11.9
pydantic_core==2.33.2
python-dotenv==1.1.1
PyYAML==6.0.3
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.48.0
typing-inspection==0.4.1
typing_extensions==4.15.0
uvicorn==0.37.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1

# Дополнительные зависимости для тестирования
httpx>=0.24.0
pytest-timeout>=2.1.0
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Зависимости для Bot API
redis>=4.5.0
asyncio-mqtt>=0.13.0
celery>=5.3.0

# Зависимости для WB API
httpx>=0.24.0
aiofiles>=23.0.0

# Зависимости для базы данных
alembic>=1.12.0
asyncpg>=0.29.0

# Зависимости для логирования и мониторинга
structlog>=23.0.0

# Зависимости для Google Sheets интеграции
google-auth>=2.23.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.100.0
